Delta time

+++ SlowAndReverb/Engine/Rendering/Graphics/AnimationSegment.cs
@@ -6,5 +6,5 @@ using System.Threading.Tasks;
 
 namespace SlowAndReverb
 {
-    public readonly record struct AnimationSegment(int Frame, int Delay);
+    public readonly record struct AnimationSegment(int Frame, float Delay);
 }
